chore: Update square_meters attribute to support decimal values
- Changed `square_meters` column in the `properties` table from `integer` to `float` to allow decimal values.
- Updated validation rules to accept numeric values instead of only integers.

+++ app/Http/Controllers/Properties/PropertyController.php
@@ -207,7 +207,7 @@ class PropertyController extends Controller
             'description' => ['required'],
             'address' => ['required'],
             'price' => ['required', 'integer'],
-            'square_meters' => ['required', 'integer'],
+            'square_meters' => ['required', 'numeric'],
             'city' => ['required'],
             'district' => ['required'],
             'bathrooms' => ['integer', 'nullable'],
